notes-backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pymysql
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic-settings
email-validator
python-multipart